Removed float32 -> int16 compression from the "craw" buffer in physiology acq
GvT noticed that there was some "clipping" of the raw waveform.  The 32 to 16
bit compression assumes that the signal will never exceed +/- 4 mV; however,
when recording the raw signal using the wired headstage, this is certainly a
possibility if there are low-frequency fluctuations in the waveform (i.e.
transient near-DC offsets).  I have not observed such artifacts in the wireless
system, probably because the highpass cutoff frequency is a couple of Hz higher.
The compression remains enabled for the "cfilt" buffer since this is used only
for on-line plotting (not to mention the highpass cutoff default of 300 Hz
guarantees the signal will not exceed +/- 4 mV).

+++ experiments/physiology_controller.py
@@ -70,7 +70,7 @@ class PhysiologyController(Controller):
 
         # Initialize the buffers that will be spooling the data
         self.buffer_raw = self.iface_physiology.get_buffer('craw',
-                'r', src_type='int16', dest_type='float32', channels=CHANNELS,
+                'r', src_type='float32', dest_type='float32', channels=CHANNELS,
                 block_size=1048)
         self.buffer_filt = self.iface_physiology.get_buffer('cfilt',
                 'r', src_type='int16', dest_type='float32', channels=CHANNELS,
